LayoutTests/platform/blackberry/editing/spelling/context-menu-suggestions-expected.txt

CONSOLE MESSAGE: line 32: TypeError: Result of expression 'itemNamesWithSpellChecking' [undefined] is not an object.
This test checks that spelling suggestions in the context menu are not show when spellcheck="false". To test manually, right click on the text in the input box.


FAIL: Did not run test
